requirements.txt
fastapi
uvicorn[standard]
websockets
pydantic
PyPDF2
pillow
pytesseract
gunicorn

